refactored horizontal alignment class, splitting it in two
