+++ freecad/trails/alignment/alignment_model.py
@@ -22,213 +22,37 @@
 #***********************************************************************
 
 """
-Class for managing 2D Horizontal Alignments
+Class for managing 2D Horizontal Alignment data
 """
-from copy import deepcopy
-
 import FreeCAD as App
-import Draft
 
-from ..project.support import properties, units, utils
+from ..project.support import units
 from ..geometry import arc, line, support
-from . import alignment_group
 
-_CLASS_NAME = 'HorizontalAlignment'
-_TYPE = 'Part::Part2DObjectPython'
+_CLASS_NAME = 'AlignmentModel'
+_TYPE = 'AlignmentModel'
 
-__title__ = 'horizontal_alignment.py'
+__title__ = 'alignment_model.py'
 __author__ = 'Joel Graff'
 __url__ = "https://www.freecadweb.org"
-
-
-def create(geometry, object_name='', no_visual=False):
-    """
-    Class construction method
-    object_name - Optional. Name of new object.  Defaults to class name.
-    no_visual - If true, generates the object without a ViewProvider.
-    """
-
-    if not geometry:
-        print('No curve geometry supplied')
-        return
-
-    _obj = None
-    _name = _CLASS_NAME
-
-    if object_name:
-        _name = object_name
-
-    parent = alignment_group.create()
-
-    _obj = parent.Object.newObject(_TYPE, _name)
-
-    result = _HorizontalAlignment(_obj, _name)
-    result.set_geometry(geometry)
-
-    if not no_visual:
-        Draft._ViewProviderWire(_obj.ViewObject)
-
-    App.ActiveDocument.recompute()
-    return result
 
 #Construction order:
 #Calc arc parameters
 #Sort arcs
 #calculate arc start coordinates by internal position and apply to arcs
 
-class _HorizontalAlignment(Draft._Wire):
-
-    def __init__(self, obj, label=''):
+class AlignmentModel:
+    """
+    Alignment model for the alignment FeaturePython class
+    """
+    def __init__(self, geometry):
         """
         Default Constructor
         """
-
-        super(_HorizontalAlignment, self).__init__(obj)
-
-        self.no_execute = True
-
-        obj.Proxy = self
-        
-        self.Type = _CLASS_NAME
-        self.Object = obj
         self.errors = []
-
-        self.curve_edges = None
-
         self.data = []
-        self.meta = {}
 
-        obj.Label = label
-        obj.Closed = False
-
-        if not label:
-            obj.Label = obj.Name
-
-        #add class properties
-
-        #metadata
-        properties.add(obj, 'String', 'ID', 'ID of alignment', '')
-        properties.add(obj, 'String', 'oID', 'Object ID', '')
-
-        properties.add(
-            obj, 'Length', 'Length', 'Alignment length', 0.0, is_read_only=True
-        )
-
-        properties.add(
-            obj, 'String', 'Description', 'Alignment description', ''
-        )
-
-        properties.add(
-            obj, 'Length', 'Start Station',
-            'Starting station of the alignment', 0.0
-        )
-
-        obj.addProperty(
-            'App::PropertyEnumeration', 'Status', 'Base', 'Alignment status'
-        ).Status = ['existing', 'proposed', 'abandoned', 'destroyed']
-
-        properties.add(obj, 'VectorList', 'Station Equations',
-                       'Station equation along the alignment', [])
-        properties.add(
-            obj, 'Vector', 'Datum', 'Datum value as Northing / Easting',
-            App.Vector(0.0, 0.0, 0.0)
-        )
-
-        properties.add(
-            obj, 'FloatList', 'Hashes', 'Coordiante hashes', []
-        )
-
-
-        #geometry
-        properties.add(
-            obj, 'VectorList', 'PIs', """
-            Discretization of Points of Intersection (PIs) as a list of
-            vectors""", []
-            )
-
-        properties.add(
-            obj, 'Link', 'Parent Alignment',
-            'Links to parent alignment object', None
-        )
-
-        subdivision_desc = """
-            Method of Curve Subdivision\n
-            \nTolerance - ensure error between segments and curve is (n)
-            \nInterval - Subdivide curve into segments of fixed length
-            \nSegment - Subdivide curve into equal-length segments
-            """
-
-        obj.addProperty(
-            'App::PropertyEnumeration', 'Method', 'Segment', subdivision_desc
-        ).Method = ['Tolerance', 'Interval', 'Segment']
-
-        properties.add(obj, 'Float', 'Segment.Seg_Value',
-                       'Set the curve segments to control accuracy', 1.0)
-
-        delattr(self, 'no_execute')
-
-    def __getstate__(self):
-        return self.Type
-
-    def __setstate__(self, state):
-        if state:
-            self.Type = state
-
-    def onDocumentRestored(self, obj):
-        """
-        Restore object references on reload
-        """
-
-        self.Object = obj
-
-        self.data = self.Object.InList[0].Proxy.get_alignment_data(obj.ID)
-
-        self.build_curve_edge_dict()
-
-    def build_curve_edge_dict(self):
-        """
-        Build the dictionary which correlates edges to their corresponding
-        curves for quick lookup when curve editing
-        """
-
-        curve_dict = {}
-        curves = self.data['geometry']
-
-        #iterate the curves, creating the dictionary for each curve
-        #that lists it's wire edges keyed by it's Edge index
-        for curve in curves:
-
-            if curve['Type'] == 'line':
-                continue
-
-            curve_edges = self.Object.Shape.Edges
-            curve_pts = [curve['Start'], curve['End']]
-            edge_dict = {}
-
-            #iterate edge list, add edges that fall within curve limits
-            for _i, edge in enumerate(curve_edges):
-
-                edge_pts = [edge.Vertexes[0].Point, edge.Vertexes[1].Point]
-
-                #empty list means we haven't found the start yet
-                if not edge_dict:
-
-                    if not support.within_tolerance(curve_pts[0], edge_pts[0]):
-                        continue
-
-                #still here?  then we're within the geometry
-                edge_dict['Edge' + str(_i + 1)] = edge
-
-                #if this edge fits the end, we're done
-                if support.within_tolerance(curve_pts[1], edge_pts[1]):
-                    break
-
-            #calculate a unique hash based on the curve start and end points
-            #and save the edge list to it
-
-            curve_dict[curve['Hash']] = edge_dict
-
-        self.curve_edges = curve_dict
+        self.construct_geometry(geometry)
 
     def get_pi_coords(self):
         """
@@ -238,56 +62,22 @@ class _HorizontalAlignment(Draft._Wire):
 
         result = [self.data['meta']['Start']]
         result += [_v['PI'] for _v in self.data['geometry'] if _v.get('PI')]
-        result.append(self.Object.Shape.Vertexes[-1].Point)
+
+        end_pt = self.data['meta'].get('End')
+
+        if end_pt is None:
+            self.discretize_geometry()
+
+        result.append(self.data['meta']['End'])
 
         return result
 
-    def get_edges(self):
-        """
-        Return the dictionary of curve edges
-        """
-
-        return self.curve_edges
-
-    def get_data(self):
-        """
-        Return the complete dataset for the alignment
-        """
-
-        return self.data
-
-    def get_data_copy(self):
-        """
-        Returns a deep copy of the alignment dataset
-        """
-
-        return deepcopy(self.data)
-
-    def get_geometry(self, curve_hash=None):
-        """
-        Return the geometry of the curve matching the specified hash
-        value.  If no match, return all of the geometry
-        """
-
-        if not curve_hash:
-            return self.data['geometry']
-
-        for _geo in self.data['geometry']:
-
-            if _geo['Hash'] == curve_hash:
-                return _geo
-
-        return None
-
-    def set_geometry(self, geometry):
+    def construct_geometry(self, geometry):
         """
         Assign geometry to the alignment object
         """
 
         self.data = geometry
-
-        self.assign_meta_data()
-        self.assign_station_data()
 
         for _i, _geo in enumerate(self.data['geometry']):
 
@@ -316,8 +106,6 @@ class _HorizontalAlignment(Draft._Wire):
 
         #call once more to catch geometry added by validate_alignment()
         self.validate_stationing()
-
-        self.build_curve_edge_dict()
 
         return True
 
@@ -600,10 +388,14 @@ class _HorizontalAlignment(Draft._Wire):
     def _get_internal_station(self, station):
         """
         Using the station equations, determine the internal station
-        (position) along the alingment, scaled to the document units
+        (position) along the alignment, scaled to the document units
         """
-        start_sta = self.Object.Station_Equations[0].y
-        eqs = self.Object.Station_Equations
+        start_sta = self.data['meta'].get('StartStation')
+
+        if not start_sta:
+            start_sta = 0.0
+
+        eqs = self.data['station']
 
         #default to distance from starting station
         position = 0.0
@@ -624,87 +416,11 @@ class _HorizontalAlignment(Draft._Wire):
 
         return position * units.scale_factor()
 
-    def assign_meta_data(self):
-        """
-        Extract the meta data for the alignment from the data set
-        Check it for errors
-        Assign properties
-        """
-
-        obj = self.Object
-
-        meta = self.data['meta']
-
-        if meta.get('ID'):
-            obj.ID = meta['ID']
-
-        if meta.get('Description'):
-            obj.Description = meta['Description']
-
-        if meta.get('ObjectID'):
-            obj.ObjectID = meta['ObjectID']
-
-        if meta.get('Length'):
-            obj.Length = meta['Length']
-
-        if meta.get('Status'):
-            obj.Status = meta['Status']
-
-        if meta.get('StartStation'):
-            obj.Start_Station = str(meta['StartStation']) + ' ft'
-
-    def assign_station_data(self):
-        """
-        Assign the station and intersection equation data
-        """
-
-        obj = self.Object
-
-        _eqs = self.data['station']
-        _meta = self.data['meta']
-
-        _eqn_list = []
-        _start_sta = utils.to_float(_meta.get('StartStation'))
-
-        if _start_sta:
-            _eqn_list.append(App.Vector(0.0, _start_sta, 0.0))
-        else:
-            self.errors.append('Unable to convert starting station %s'
-                               % _meta.get('StartStation')
-                              )
-
-        for _eqn in _eqs:
-
-            eq_vec = None
-
-            #default to increasing station if unspecified
-            if not _eqn['Direction']:
-                _eqn['Direction'] = 1.0
-
-            try:
-                eq_vec = App.Vector(
-                    float(_eqn['Back']),
-                    float(_eqn['Ahead']),
-                    float(_eqn['Direction'])
-                    )
-
-            except ValueError:
-                self.errors.append(
-                    'Unable to convert station equation (%s)' % (_eqn)
-                )
-                continue
-
-            _eqn_list.append(eq_vec)
-
-        obj.Station_Equations = _eqn_list
-
-    def discretize_geometry(self):
+    def discretize_geometry(self, interval=10.0, interval_type='Segment'):
         """
         Discretizes the alignment geometry to a series of vector points
         """
 
-        interval = self.Object.Seg_Value
-        interval_type = self.Object.Method
         geometry = self.data['geometry']
         points = [[self.data['meta']['Start']]]
         last_curve = None
@@ -762,95 +478,7 @@ class _HorizontalAlignment(Draft._Wire):
 
             result.append(last_point.add(_vec))
 
+        if not self.data['meta'].get('End'):
+            self.data['meta']['End'] = result[-1]
+
         return result
-
-    def onChanged(self, obj, prop):
-        """
-        Property change callback
-        """
-        #dodge onChanged calls during initialization
-        if hasattr(self, 'no_execute'):
-            return
-
-        if prop == "Method":
-
-            _prop = obj.getPropertyByName(prop)
-
-            if _prop == 'Interval':
-                self.Object.Seg_Value = 100.0
-
-            elif _prop == 'Segment':
-                self.Object.Seg_Value = 10.0
-
-            elif _prop == 'Tolerance':
-                self.Object.Seg_Value = 1.0
-
-    def execute(self, obj):
-        """
-        Recompute callback
-        """
-        if hasattr(self, 'no_execute'):
-            return
-
-        points = self.discretize_geometry()
-
-        if not points:
-            return
-
-        self.Object.Points = points
-
-        super(_HorizontalAlignment, self).execute(obj)
-        #self.Object.Placement.Base = self.Object.Placement.Base.add
-        #(self.get_intersection_delta())
-        #super(_HorizontalAlignment, self).execute(obj)
-
-class _ViewProviderHorizontalAlignment:
-
-    def __init__(self, obj):
-        """
-        Initialize the view provider
-        """
-        obj.Proxy = self
-        self.Object = obj
-
-    def __getstate__(self):
-        return None
-
-    def __setstate__(self, state):
-        return None
-
-    def attach(self, obj):
-        """
-        View provider scene graph initialization
-        """
-        self.Object = obj
-
-    def updateData(self, obj, prop):
-        """
-        Property update handler
-        """
-        pass
-
-    def getDisplayMode(self, obj):
-        """
-        Valid display modes
-        """
-        return ["Wireframe"]
-
-    def getDefaultDisplayMode(self):
-        """
-        Return default display mode
-        """
-        return "Wireframe"
-
-    def setDisplayMode(self, mode):
-        """
-        Set mode - wireframe only
-        """
-        return "Wireframe"
-
-    def onChanged(self, vobj, prop):
-        """
-        Handle individual property changes
-        """
-        pass
